Replace Footer w/ SectionWithFooter
