+++ frontend/svelte/src/routes/Proposals.svelte
@@ -3,7 +3,7 @@
   import { onMount } from "svelte";
   import VotingFilters from "../lib/components/VotingFilters.svelte";
   import { i18n } from "../lib/stores/i18n";
-  import Footer from "../lib/components/Footer.svelte";
+  import SectionWithFooter from "../lib/components/SectionWithFooter.svelte";
 
   // TODO: To be removed once this page has been implemented
   onMount(() => {
@@ -15,7 +15,7 @@
 
 {#if !process.env.REDIRECT_TO_LEGACY}
   <Layout>
-    <section>
+    <SectionWithFooter>
       <h1>{$i18n.voting.title}</h1>
 
       <p>{$i18n.voting.text}</p>
@@ -23,7 +23,6 @@
       <VotingFilters />
 
       <!-- TODO(#L2-206): list proposals and use filters -->
-      <Footer />
-    </section>
+    </SectionWithFooter>
   </Layout>
 {/if}
